Update assign1.Rmd
Switched around the code that created the "results" data frame, to before the creation of the map. This way the lines on the map that show the dissimilarity index can be drawn from the values in this data frame. From the results df, I then created a vector that exponentiated all the values, and then had the lines drawn from positions in the vector, rather than having to hard code in the numbers, since this is amenable now to changing values if other analysis' or data sets were wanting to be explored.