Update map plot
Wanted to change the line colours based on what the dissimilarity index values were for each line between two points. Just to illustrate the difference in values between different points a little better. To do this used the function I created, "col.fun", that was sourced in a previous commit. Then ran this function inside the "col" function in the lines portions of the code for the map (Lines 227-238). Then just ran the subsetted results dataframe through the function which would print out a specific colour string depending on the dissimilarity index value. Then just changed the map legend to accomdate the colour differences as well.